Gregtech Config Settings Update
Turned on a variety of gregtech hardmore recipes. Namely gregified recipes of certain items. Of greatest interest will be that crafting rods in an extruder is now more efficient than in a lathe and recipe times reset if the machine loses power (no more playing on and off with the machine's power).

There may be cleanup to occur over time to bring recipes from other mods in line with the gregification of vanilla items such as doors, trapdoors, bookshelves, etc.